# workshop-api/openapi.yaml
openapi: "3.0.0"
info:
  title: HTL-Workshop API
  version: 1.0.0
  description: >-
    The API based on the lab-description
  contact:
    name: Cloudflight
    url: 'https://cloudflight.io'
servers:
  - url: /api

paths:
  /employees:
    description: Receive a list of all employees
    get:
      operationId: listEmployees
      tags:
        - employee
      responses:
        '200':
          description: An array of all exisiting employees
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EmployeeListEntryDTO"
    post:
      summary: fetches list containing employees filtered by partial name
      operationId: listFilteredEmployees
      tags:
        - employee
      parameters:
        - name: partialName
          in: query
          schema:
            type: string
      responses:
        '200':
          description: An array of existing employees containing the partial name
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EmployeeListEntryDTO"

  /employee/{id}:
    description: Receive details about a single employee
    get:
      operationId: findEmployeeDetails
      tags:
        - employee
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Details about an exisiting employee.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmployeeWithTasksDTO"
        '404':
          description: No employee was found with the given id.

components:
  schemas:
    EmployeeListEntryDTO:
      required:
        - id
        - firstName
        - lastName
      properties:
        id:
          type: string
          description: "The identifier of a single employee"
        firstName:
          type: string
        lastName:
          type: string
    EmployeeWithTasksDTO:
      required:
        - id
        - firstName
        - lastName
        - tasks
      properties:
        id:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        tasks:
          type: array
          items:
            $ref: "#/components/schemas/TaskDto"
    TaskDto:
      required:
        - id
        - description
        - finished
        - hoursWorked
      properties:
        id:
          type: integer
          format: int32
        description:
          type: string
        finished:
          type: string
          format: date
        hoursWorked:
          type: integer
          format: int32